+++ src/main/groovy/life/qbic/datamodel/people/Person.groovy
@@ -1,6 +1,5 @@
 package life.qbic.datamodel.people;
 
-@Deprecated
 class Person {
 
   private String firstName